Create project structure with pytest as dependency
Some decisions were taken here:
- Use poetry as python packaging and dependency management
- Use pytest for testing

Poetry was chosen, because of the fun of using something I haven't used
before. This is not a serious project and my goal (besides getting the
job (done)) is to learn a couple new tools. I've came across poetry a
couple of times and never used it. So now is a good moment to try it
out.

Pytest was chosen because I know it very well and writing tests in it is
a bliss.