Expect sloper pocket and crimp gestures not to exist
I totally love the `xfail` mark in pytest... for the following reason:

At this stage of development we don't have defined any gestures yet. So
we expect the test to fail. This is pretty obvious and well known for
most software engineers.

But now comes the clue... if we commit the test as it is, without any
implementation, the testrun will be red and therefore further CI, CD
etc. will fail.

In most cases, software engineers will start implementing the
functionality, until they get the test to green (and therefore the
testrun green as well)... but how do you know that the test is testing
the right thing? What if the test never failed? How do you make sure the
test really works?

This is where xfail shines! xfail will make a testrun go green if the
marked test failed and the test itself can be committed on its own,
making sure that it runs individually through CI. Isn't this cool? I
totally love this!