Add an Enum to define the gestures of the game
There can't be many gestures in the game... as the number of connections
(c) needed to the number of possible gestures (n) is given by:

  c = n * (n - 1) / 2

This is of the order of n^2

Assuming we introduce 20 gestures (which is an acceptable number for an
enum) we would get 200 connections.

I doubt there would be many humans left capable of playing this game
with 20 gestures without looking up the rules of the game all the time.
The game becomes therefore unplayable before enum becomes a bad
structure to store the gestures.