Test gesture existance and remove import test
The previous test was removed, since the import is tested explicitely.
If for some reason `Gestures` could not be imported from
`sloper_pocket_crimp`, there would be an ERROR while running the tests.
The ERROR indicates that the testrun is failed, but that the reason is
not a failing test, but rather an error while setting up the tests. In
my opinion this is tested well enough, since `Gestures` is an
implementation detail and should be considered a private Class (a
concept which does not exist in Python, ... altough there is a the
convention of using underscores to indicate private attributes). In the
case of a compiled programming language, this is catched by the
compiler. In the case of Python, this error is raised at runtime.

To test the existance of the gestures a parametrized test is used
instead of using a for-loop. The reason is obvious. Assuming we
implemented `SLOPER` and `CRIMP`, in the case of a for-loop within the
test, we would see a failing test for `POCKET` and we wouldn't see the
test result for `CRIMP`. In the case of a parametrized test, each
attribute is tested individually giving much more information about the
state of implementation.