Add outcomes enum
Using the same arguments as for the gestures, there can't be a huge
number of outcomes, otherwise the game becomes unplayable. On the other
hand... the number of outcomes (o) needed to describe a game with a
number players (p) is given by:

  o = p * (p - 1) / 2

If business decides to evolve the game into a multiplayer game with
dozens of players playing at the same time, then enum is not the right
data structure to solve the problem and the design needs to be changed.