Add outcomes of a game
And now we're living dangerously:
We don't want the values of the enum to be used. It would be nice if
there was an Enum data structure without values in Python, but I don't
know any (there probably is one in some library somewhere, but that's
not the point here). By chosing these values, we're kind of enabling
developers to use the values to determine the player who won. This is
risky, since using these values would constraint the development of the
software. Imagine these values are used in development, then all of a
suddden we would have literals spread in our code, which need to be
changed whenever we chose to change this enum. So altough this is a bad
choice here, I'm chosing to choose this choice for the sake of having
this discussion.