Test outcome of game method with known gestures
If we assume, that sloper-pocket-crimp is a library, then game would
probably be part of the interface. In that case, these tests would be
acceptance tests. I assume sloper-pocket-crimp to be a runable, which
won't be used in other software, and therefore won't see game as part of
the interface. I therfore don't mark these tests as acceptance tests.