Verify that game raises InvalidGesture
This test is merely for showcase purposes. It's not complete, since
we're only testing if the second gesture is invalid and neglect the
first gesture. We also use very specific cases for invalid input, and
more corner and edge-cases should be tested. A testing library like
hypothesis (https://hypothesis.readthedocs.io/en/latest/) could be
helpful here.

What I want to showcase here is fixtures and their parametrization. This
is quite an awesome pytest feature I totally love.