Verify if human player returns a choice
...this time I'll skip the import test, since we've done it more than
enough and it's going to be removed anyway. I add the few lines of code
necessary for the import to pass and write the test which verifies that
a choice given by the user is a valid gesture.

Since we're going to use the CLI, we're going to use the builtin `input`
function, which asks the user to type in some value. We therefore patch
this builtin method to return a valid value. Invalid values are going to
be tested in a later step.