Stabilize listing available gestures
We wrote a spike in the previous commit to identify the needed methods.
Now that we know what they are, we can write the tests to make them
behave the way we expect them. Here I write a test to make sure that the
print output listing the available gestures is appealing. I expect it to
fail, since I didn't  work on the formatting of the output at all... and
indeed it does fail.

Next I can improve the method listing the available methods let the test
pass.