Stabilize chose gesture method
...and remove the previous test testing the choice method.

The behavior of choice is going to be tested in a more detailed way in
upcoming tests. The newly added test is mostly a replica of the removed
test.

We could argue if we should remove the choice method in this commit, as
it is untested and to some extent I would aggree, since it is not used
anywhere else within the project. But in practice we will deal with many
functions which are not tested and which cannot be removed. I keep it
therefore as a showcase for further stabilization.