Change the choice interface
...by passing it the list of available gestures.

This allows to get rid of the global variable and makes the code
cleaner and can be tested better, since there is no need to patch the
gestures in the tests.

It's not that we would need it in this case, but for me this project is
mostly about having fun and having a base for discussion of technical
skills. We need to see controversy to get to good discussions.