Verify that choice can handle bad user input
Ok... first things first. This test is bad. And it is bad for the sake
of having a discussion. What's bad about the test? Quite a lot of
things:
- It tests more than one thing at the time
- It tests knowing the implementation details (I prefer opaque box tests
  over transparent box tests)
- There is implicit data in the test (it's not clear that 0, 1 and 2 are
  the indices for the expected gestures

I wouldn't write this test this way but rather write two separate tests.
One to test that the selected gesture is a valid gesture, then one for
the number of iterations until the answer was valid. I would also use a
library like hypothesis to generate invalid user input.

Now looking at the bright and dangerous side of the test:
This is the first time we use a MagicMock within this project. Magic
mocks are fun and helpful, but on the other hand they're a smell.
Overusing MagicMock can lead to tests not testing the software but
rather the test setup. As mentioned before, I prefer having opaque box
tests rather than transparent box tests. Is it just a matter of taste?