Refactor game into a class
You can't see it here, since it's just one commit, but I adapted first
the tests, adding a game fixture and then moved on to adapt the code.

The idea behind the refactor is to move all the logic and state into a
single object.

In the next steps we will need to add logic to instantiate the player
depending on the play mode (computer vs. human or computer vs.
computer), then letting each player select their gesture and display the
outcome.