Implement player generation
There is again room for discussion within this change. We moved the
definition of the GameModes from the test code to production code. I had
an interesting discussion about this topic with an ex coworker of mine,
we noticed how often, we start defining behavior and designing classes
and functions in the test code, then, as soon at it is defined well
enough, this code is needed within production.